requirements.txt
asgiref==3.6.0
attrs==22.2.0
backports.zoneinfo;python_version<"3.9"
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==39.0.2
defusedxml==0.7.1
Django==4.1.7
django-cors-headers==3.13.0
django-filter==22.1
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-spectacular==0.26.0
drf-yasg==1.21.5
idna==3.4
importlib-metadata==6.0.0
importlib-resources==5.12.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.3
Markdown==3.4.1
MarkupSafe==2.1.2
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.0
Pillow==9.4.0
pkgutil-resolve-name==1.3.10
psycopg2-binary==2.9.5
pycparser==2.21
Pygments==2.14.0
PyJWT==2.6.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-decouple==3.7
python3-openid==3.2.0
pytz==2022.7.1
PyYAML==6.0
requests==2.28.2
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.4
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.4.0
sqlparse==0.4.3
swagger-spec-validator==3.0.3
typing-extensions==4.5.0
uritemplate==4.1.1
urllib3==1.26.14
zipp==3.14.0
